+++ packages/plugin-swr/src/components/Mutation.tsx
@@ -210,12 +210,10 @@ Mutation.File = function ({ templates = defaultTemplates }: FileProps): ReactNod
     },
   } = useApp<PluginSwr>()
 
-  const { getSchemas, getFile, getName } = useOperationManager()
+  const { getFile, getName } = useOperationManager()
   const operation = useOperation()
 
-  const schemas = getSchemas(operation, { pluginKey: [pluginTsName], type: 'type' })
   const file = getFile(operation)
-  const fileType = getFile(operation, { pluginKey: [pluginTsName] })
   const factoryName = getName(operation, { type: 'type' })
 
   const Template = templates.default
@@ -228,19 +226,6 @@ Mutation.File = function ({ templates = defaultTemplates }: FileProps): ReactNod
       <File.Import name="useSWRMutation" path="swr/mutation" />
       <File.Import name={['SWRMutationConfiguration', 'SWRMutationResponse']} path="swr/mutation" isTypeOnly />
       <File.Import name={'client'} path={importPath} />
-      <File.Import
-        name={[
-          schemas.request?.name,
-          schemas.response.name,
-          schemas.pathParams?.name,
-          schemas.queryParams?.name,
-          schemas.headerParams?.name,
-          ...(schemas.errors?.map((error) => error.name) || []),
-        ].filter(Boolean)}
-        root={file.path}
-        path={fileType.path}
-        isTypeOnly
-      />
 
       <SchemaType factory={factory} />
       <Mutation Template={Template} factory={factory} />
