+++ packages/plugin-tanstack-query/src/components/SchemaType.tsx
@@ -17,10 +17,12 @@ export function SchemaType({ factory }: Props): ReactNode {
       options: { dataReturnType },
     },
   } = useApp<PluginTanstackQuery>()
-  const { getSchemas } = useOperationManager()
+  const { getSchemas, getFile } = useOperationManager()
   const operation = useOperation()
 
   const schemas = getSchemas(operation, { pluginKey: [pluginTsName], type: 'type' })
+  const file = getFile(operation)
+  const fileType = getFile(operation, { pluginKey: [pluginTsName] })
 
   const [TData, TError, TRequest, TPathParams, TQueryParams, THeaderParams, TResponse] = [
     schemas.response.name,
@@ -37,6 +39,19 @@ export function SchemaType({ factory }: Props): ReactNode {
 
   return (
     <>
+      <File.Import
+        name={[
+          schemas.request?.name,
+          schemas.response.name,
+          schemas.pathParams?.name,
+          schemas.queryParams?.name,
+          schemas.headerParams?.name,
+          ...(schemas.errors?.map((error) => error.name) || []),
+        ].filter(Boolean)}
+        root={file.path}
+        path={fileType.path}
+        isTypeOnly
+      />
       <File.Source name={clientType} isTypeOnly>
         <Type name={clientType}>{`typeof client<${TResponse}, ${TError}, ${isFormData ? 'FormData' : TRequest}>`}</Type>
       </File.Source>
